# backend_starter_skeleton_v5/api/openapi.yaml
openapi: 3.0.3
info:
  title: Rental Mobil API
  version: "1.0.0"
  description: >
    MVP Backend API untuk Platform Rental Mobil (Admin/Owner). Semua timestamp UTC ISO-8601.
servers:
  - url: /v1
...
paths:
...
  /vehicles:
    get:
      tags: [vehicles]
      security: [{ bearerAuth: [] }]
      summary: List vehicles
      parameters:
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/VehicleStatus" }
        - in: query
          name: is_active
          schema: { type: boolean }
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedVehicles"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [vehicles]
      security: [{ bearerAuth: [] }]
      summary: Create vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer, example: 1 }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /vehicles/{id}:
    get:
      tags: [vehicles]
      security: [{ bearerAuth: [] }]
      summary: Get vehicle detail
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [vehicles]
      security: [{ bearerAuth: [] }]
      summary: Update vehicle
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /vehicles/{id}/status:
    patch:
      tags: [vehicles]
      security: [{ bearerAuth: [] }]
      summary: Patch vehicle status / active (owner-only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchVehicleStatusRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unauthorized:
              value:
                error: { code: UNAUTHORIZED, message: "Token tidak valid atau sudah expired.", details: null }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            forbidden:
              value:
                error: { code: FORBIDDEN, message: "Anda tidak memiliki akses.", details: null }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            notFound:
              value:
                error: { code: NOT_FOUND, message: "Resource tidak ditemukan.", details: null }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            validation:
              value:
                error: { code: VALIDATION_ERROR, message: "Validasi gagal.", details: { field: "end_at", reason: "must be after start_at" } }
    Conflict:
      description: Conflict (business rule)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            overlap:
              value:
                error: { code: RENTAL_OVERLAP, message: "Kendaraan sudah disewa pada rentang waktu tersebut.", details: null }
...
  schemas:
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    VehicleStatus:
      type: string
      enum: [AVAILABLE, RENTED, MAINTENANCE, INACTIVE]
    Vehicle:
      type: object
      required: [id, name, plate_number, type, price_per_day, status, is_active, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        plate_number: { type: string }
        type: { type: string }
        price_per_day: { type: integer }
        status: { $ref: "#/components/schemas/VehicleStatus" }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time, nullable: true }
    CreateVehicleRequest:
      type: object
      required: [name, plate_number, type, price_per_day]
      properties:
        name: { type: string }
        plate_number: { type: string }
        type: { type: string }
        price_per_day: { type: integer, minimum: 0 }
        is_active: { type: boolean, default: true }
    UpdateVehicleRequest:
      type: object
      properties:
        name: { type: string }
        plate_number: { type: string }
        type: { type: string }
        price_per_day: { type: integer, minimum: 0 }
        is_active: { type: boolean }
    PatchVehicleStatusRequest:
      type: object
      properties:
        status: { $ref: "#/components/schemas/VehicleStatus" }
        is_active: { type: boolean }
    PagedVehicles:
      type: object
      required: [data, limit, offset]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Vehicle" }
        limit: { type: integer }
        offset: { type: integer }
...